Update olcConsoleGameEngine.h
Changed the third argument to Draw() function from type "short" to type "wchar_t". Overriding this function did not work correctly in the Asteroids tutorial when walking through it, as when defining the overridden function a type "wchar_t" was used. Therefore this broke the WrapCoordinates function applying the wrapping to each pixel being drawn.